+++ src/com/payment/Applicable.java
@@ -6,7 +6,7 @@ public interface Applicable {
     public boolean isCustomerEligible(Member member);
     public boolean isMinimumPriceEligible(Order order);
     public boolean isShippingDiscountEligible(Order order);
-    public int totalDiscount();
-    public int totalCashback();
-    public int calculateShippingDiscount();
+    public long totalDiscount();
+    public long totalCashback();
+    public long calculateShippingDiscount();
 }
